refactor : Cars 객체 생성
