+++ src/main/java/racingcar/RacingGameApplication.java
@@ -1,5 +1,6 @@
 package racingcar;
 
+import racingcar.domain.Cars;
 import racingcar.domain.RacingGame;
 import racingcar.view.InputView;
 import racingcar.view.ResultView;
@@ -10,7 +11,8 @@ public class RacingGameApplication {
         ResultView resultView = new ResultView();
 
         inputView.input();
-        RacingGame racingGame = RacingGame.newRacingGame(inputView.getCarNames(), inputView.getRounds());
+        Cars cars = new Cars(inputView.getCarNames());
+        RacingGame racingGame = RacingGame.newRacingGame(cars, inputView.getRounds());
         racingGame.play();
         resultView.print(racingGame);
     }
