+++ src/main/java/racingcar/domain/RaceRecords.java
@@ -1,0 +1,4 @@
+package racingcar.domain;
+
+public class RaceRecords {
+}
